Pushing updated InventoryController file
