+++ WMS/src/controller/.gitignore
@@ -1,0 +1,5 @@
+/Driver.class
+/InventoryController.class
+/JDBCTest.class
+/WarehouseController.class
+/JDBCtest.class
